requirements.txt
fillpdf==0.7.2
pdf2image==1.16.3
pdfrw2==0.5.0
Pillow==10.0.0
PyMuPDF==1.22.5
PyMySQL==1.1.0
python-dotenv==1.0.0
